[hotfix] Fix FlinkActionsE2eTest testDropPartition timeout (#2316)
