+++ paimon-e2e-tests/src/test/java/org/apache/paimon/tests/FlinkActionsE2eTest.java
@@ -18,124 +18,18 @@
 
 package org.apache.paimon.tests;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.testcontainers.containers.Container;
 
-import java.util.UUID;
-
 /** Tests for {@code FlinkActions}. */
-public class FlinkActionsE2eTest extends E2eTestBase {
+public class FlinkActionsE2eTest extends FlinkActionsE2eTestBase {
 
     private static final Logger LOG = LoggerFactory.getLogger(FlinkActionsE2eTest.class);
 
     public FlinkActionsE2eTest() {
-        super(true, false);
-    }
-
-    private String warehousePath;
-    private String catalogDdl;
-    private String useCatalogCmd;
-
-    @BeforeEach
-    public void setUp() {
-        warehousePath = TEST_DATA_DIR + "/" + UUID.randomUUID() + ".store";
-        catalogDdl =
-                String.format(
-                        "CREATE CATALOG ts_catalog WITH (\n"
-                                + "    'type' = 'paimon',\n"
-                                + "    'warehouse' = '%s'\n"
-                                + ");",
-                        warehousePath);
-
-        useCatalogCmd = "USE CATALOG ts_catalog;";
-    }
-
-    @Test
-    public void testCompact() throws Exception {
-        String topicName = "ts-topic-" + UUID.randomUUID();
-        createKafkaTopic(topicName, 1);
-        // prepare first part of test data
-        sendKafkaMessage("1.csv", "20221205,1,100\n20221206,1,100\n20221207,1,100", topicName);
-
-        String testDataSourceDdl =
-                String.format(
-                        "CREATE TEMPORARY TABLE test_source (\n"
-                                + "    dt STRING,\n"
-                                + "    k INT,\n"
-                                + "    v INT"
-                                + ") WITH (\n"
-                                + "    'connector' = 'kafka',\n"
-                                + "    'properties.bootstrap.servers' = 'kafka:9092',\n"
-                                + "    'properties.group.id' = 'testGroup',\n"
-                                + "    'scan.startup.mode' = 'earliest-offset',\n"
-                                + "    'topic' = '%s',\n"
-                                + "    'format' = 'csv'\n"
-                                + ");",
-                        topicName);
-
-        String tableDdl =
-                "CREATE TABLE IF NOT EXISTS ts_table (\n"
-                        + "    dt STRING,\n"
-                        + "    k INT,\n"
-                        + "    v INT,\n"
-                        + "    PRIMARY KEY (dt, k) NOT ENFORCED\n"
-                        + ") PARTITIONED BY (dt) WITH (\n"
-                        + "    'changelog-producer' = 'full-compaction',\n"
-                        + "    'changelog-producer.compaction-interval' = '1s',\n"
-                        + "    'write-only' = 'true'\n"
-                        + ");";
-
-        // insert data into paimon
-        runSql(
-                "SET 'execution.checkpointing.interval' = '1s';\n"
-                        + "INSERT INTO ts_table SELECT * FROM test_source;",
-                catalogDdl,
-                useCatalogCmd,
-                tableDdl,
-                testDataSourceDdl);
-
-        // run dedicated compact job
-        Container.ExecResult execResult =
-                jobManager.execInContainer(
-                        "bin/flink",
-                        "run",
-                        "-D",
-                        "execution.checkpointing.interval=1s",
-                        "--detached",
-                        "lib/paimon-flink-action.jar",
-                        "compact",
-                        "--warehouse",
-                        warehousePath,
-                        "--database",
-                        "default",
-                        "--table",
-                        "ts_table",
-                        "--partition",
-                        "dt=20221205",
-                        "--partition",
-                        "dt=20221206");
-        LOG.info(execResult.getStdout());
-        LOG.info(execResult.getStderr());
-
-        // read all data from paimon
-        runSql(
-                "INSERT INTO result1 SELECT * FROM ts_table;",
-                catalogDdl,
-                useCatalogCmd,
-                tableDdl,
-                createResultSink("result1", "dt STRING, k INT, v INT"));
-
-        // check that first part of test data are compacted
-        checkResult("20221205, 1, 100", "20221206, 1, 100");
-
-        // prepare second part of test data
-        sendKafkaMessage("2.csv", "20221205,1,101\n20221206,1,101\n20221207,1,101", topicName);
-
-        // check that second part of test data are compacted
-        checkResult("20221205, 1, 101", "20221206, 1, 101");
+        super(false, false);
     }
 
     @Test
@@ -400,7 +294,7 @@ public class FlinkActionsE2eTest extends E2eTestBase {
         checkResult();
     }
 
-    private void runSql(String sql, String... ddls) throws Exception {
+    protected void runSql(String sql, String... ddls) throws Exception {
         runSql(String.join("\n", ddls) + "\n" + sql);
     }
 }
